Bug fixes and balancing changes related to the UI and to the scale of the game pieces

+++ Assets/Scripts/Save/SaveData.cs
@@ -6,8 +6,8 @@ using System.Collections.Generic;
 [System.Serializable]
 public class SaveData 
 {
-    public int highScore = 0;
-    public int highChain = 0;
+    public float highScore = 0;
+    public float highChain = 0;
     public bool SFXOn = true;
     public bool musicOn = true;
 }
